Rename workflow text block to content
