+++ client/web/workflow/src/components/Configurator/loader.js
@@ -10,4 +10,4 @@ export { default as Prompt } from './Prompt'
 export { default as Delay } from './Delay'
 export { default as ExecWorkflow } from './ExecWorkflow'
 export { default as ErrorHandler } from './ErrorHandler'
-export { default as Text } from './Text'
+export { default as Content } from './Content'
